split web related implementation to epi_web.rs
